requirements.txt
attrs==21.2.0
genson==1.2.2
Jinja2==3.0.1
jsonschema==3.2.0
pyyaml==5.4.1
rfc3987==1.3.8